+++ assets/js/global-admin.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array(), 'version' => 'b756bc35b4f8fd57f1d2');
+<?php return array('dependencies' => array(), 'version' => '5c68106cf1cbc29414b2');