+++ includes/Admin/SetupWizard.php
@@ -506,7 +506,7 @@ class SetupWizard {
     /**
      * Commission step.
      *
-     * @since DOKAN_SINCE
+     * @since 3.14.5
      *
      * @return void
      */
@@ -553,7 +553,7 @@ class SetupWizard {
     /**
      * Save commission options.
      *
-     * @since DOKAN_SINCE
+     * @since 3.14.5
      *
      * @return void
      */