+++ tests/pw/tests/api/calculation.spec.ts
@@ -26,11 +26,11 @@ test.describe.serial('commission calculation test', () => {
         await apiUtils.dispose();
     });
 
-    test.skip('calculation (debug)', { tag: ['@lite'] }, async () => {
-        const orderId = '4100';
-        const [orderResponse, orderResponseBody] = await apiUtils.getSingleOrder(orderId);
-        await assertOrderCalculation([orderResponse, orderResponseBody, orderId]);
-    });
+    // test.skip('calculation (debug)', { tag: ['@lite'] }, async () => {
+    //     const orderId = '4100';
+    //     const [orderResponse, orderResponseBody] = await apiUtils.getSingleOrder(orderId);
+    //     await assertOrderCalculation([orderResponse, orderResponseBody, orderId]);
+    // });
 
     test('global commission fixed (only percentage)', { tag: ['@lite'] }, async () => {
         // set order condition